+++ WeatherForecast/WeatherForecast/Models/LocationManager.swift
