+++ WeatherForecast/WeatherForecast/Models/LocationManagerDelegate.swift
