+++ WeatherForecast/WeatherForecast/Service/Manager/OpenWeatherSessionDelegate.swift
