+++ WeatherForecast/WeatherForecast/View/Delegate/EmptyDelegate.swift
